feat: fixes an off-by-one causing a missing item, and improved randomness
The C library needed to use a re-entrant version of rand() in order to
not deterministically generate shuffled arrays.

Unrelated (but included in this commit, oh well), the schema now
defaults to "full_name" instead of "name", as the generation script has
also changed to now do first, last, and/or full names.